+++ source/TestCommon/source/FunctionApp.TestCommon/ServiceBus/ListenerMock/ServiceBusListenerMock.cs
@@ -164,9 +164,9 @@ namespace Energinet.DataHub.Core.FunctionApp.TestCommon.ServiceBus.ListenerMock
         /// Reset all listeners.
         /// </summary>
         /// <remarks>Avoid using this, unless you are removing the queue or topic</remarks>
-        public async Task ResetMessageReceiversAsync()
+        public Task ResetMessageReceiversAsync()
         {
-            await CleanupMessageReceiversAsync().ConfigureAwait(false);
+            return CleanupMessageReceiversAsync();
         }
 
         /// <summary>
@@ -278,7 +278,12 @@ namespace Energinet.DataHub.Core.FunctionApp.TestCommon.ServiceBus.ListenerMock
             foreach (var registration in MessageReceivers)
             {
                 var receiver = registration.Value;
+
                 await receiver.StopProcessingAsync().ConfigureAwait(false);
+
+                receiver.ProcessMessageAsync -= HandleMessageReceivedAsync;
+                receiver.ProcessErrorAsync -= HandleMessagePumpExceptionAsync;
+
                 await receiver.DisposeAsync().ConfigureAwait(false);
             }