+++ source/TestCommon/documents/release-notes/release-notes.md
@@ -1,5 +1,10 @@
 # TestCommon Release notes
 
+## Version 1.0.0
+
+- Implemented `ServiceBusResourceProvider`. See [Resource providers](../functionapp-testcommon.md#resource-providers).
+- Implemented `IntegrationTestConfiguration` to support using Integration Test environment. See [Integration Test environment](../functionapp-testcommon.md#integration-test-environment).
+
 ## Version 0.0.1
 
 - Preparing packages for initial release.